Update form_collections.rst
The text above is talking about the many-to-many relationship and the Yaml + XML is showing one-to-many options. Also I've dug up doctrine and they use the "cascade" persist with a double quotes around "persist".

I've followed all these steps and the one-to-many Form Collections is still not adding my collection items. (in the addItem, I've included the "$item->setInverse($this)" as well). Not sure what's left to do. I've built a system that takes care of all my entities at the same time so I don't want to build a custom "editAction" in my controller to "edit" my "Inverse" entity. (P.S. Names aren't the real ones)